Move remaining CoreRT/ProjectN code to shared partition
